Update current value as per pace.

+++ include/ObdInfo.hpp
@@ -11,7 +11,7 @@ class ObdInfo
 {
 public:
     ObdInfo(uint16_t pid, uint8_t length, std::string description, double current,
-            double min, double max, double increment, int pace,
+            double min, double max, double increment, uint64_t pace,
             std::function<int32_t(const double &current)> formulaGetter = nullptr);
 
     void update();
@@ -28,12 +28,12 @@ public:
     double getDefaultMin() const;
     double getDefaultMax() const;
     double getIncrement() const;
-    int getPace() const;
+    uint64_t getPace() const;
 
     void setMin(double min);
     void setMax(double max);
     void setIncrement(double increment);
-    void setPace(int pace);
+    void setPace(uint64_t pace);
 
 private:
     uint16_t m_pid;
@@ -47,7 +47,7 @@ private:
     double m_max;
     double m_increment;
 
-    int m_pace;
+    uint64_t m_pace;
     bool m_increasing;
     std::function<int32_t(const double &)> m_formulaGetter;
 };
